populate-depot: Don't unlink empty files and symlinks from SDK sysroot
When we're minimizing the Platform runtime, we want to delete any file
that is fully represented by its entry in the mtree, but when we're
generating the top-level mtree, we can't do that: if there was an empty
file or symlink that's functionally necessary to run pressure-vessel,
we have to preserve it, so that we can run pressure-vessel and
reconstitute the Platform from the mtree manifest.

This was particularly bad for the non-default option that adds a SDK
sysroot to the container runtime, which we don't do in production but
we do use during CI.

steamrt/tasks#448

Fixes: d78e1bbe "populate-depot: Rename files with multiple links or problematic names"

+++ subprojects/container-runtime/populate-depot.py
@@ -1487,14 +1487,17 @@ class Main:
                             rename[hashed_name] = name
                             unlink_later.add(name)
                         # else represent ./foo/bar as ./foo/bar
-                    else:
+                    elif can_rename_files:
                         unlink_later.add(name)
 
                 elif stat.S_ISLNK(stat_info.st_mode):
                     fields.append('type=link')
                     fields.append(
                         f'link={self.octal_escape(os.readlink(member))}')
-                    unlink_later.add(name)
+
+                    if can_rename_files:
+                        unlink_later.add(name)
+
                 elif stat.S_ISDIR(stat_info.st_mode):
                     fields.append('type=dir')
                 else:
@@ -1521,13 +1524,17 @@ class Main:
                 for name in sorted(not_windows_friendly):
                     writer.write('# {}\n'.format(self.octal_escape(name)))
 
-        for name, original in rename.items():
-            (top / name).parent.mkdir(parents=True, exist_ok=True)
-            (top / original).replace(top / name)
+        if can_rename_files:
+            for name, original in rename.items():
+                (top / name).parent.mkdir(parents=True, exist_ok=True)
+                (top / original).replace(top / name)
 
-        for name in unlink_later:
-            with suppress(FileNotFoundError):
-                (top / name).unlink()
+            for name in unlink_later:
+                with suppress(FileNotFoundError):
+                    (top / name).unlink()
+        else:
+            assert not rename, rename
+            assert not unlink_later, unlink_later
 
         return lc_names
